gve: Compute Engine virtual ethernet driver.
This is a linux driver for the Google Compute Engine virtual ethernet
device that supports a single Tx queue and a single Rx queue.

The google/ directory contains the upstream Kconfig and Makefile
for a google/ subdir in the kernel. It also contains a patch file
that shows what to add to the drivers/net/ethernet Kconfig and
Makefiles and some example text to use when sending these upstream.

The google/gve/ directory contains source for the driver that works
against the upstream net-next tree as well as an upstream
Makefile.

The patches/ directory contains the coccinelle patches used by
build_src.sh to create a backported driver.

LICENSE, Makefile.oot, README.md, and dkms.conf are used by build_src.sh
to build an out of tree module package that will compile against all
supported kernels.

Signed-off-by: Jon Olson <jonolson@google.com>
Signed-off-by: Catherine Sullivan <csully@google.com>