gve: Tx iovec: Remove iov_base pointer, saves 32 bytes/tx queue entry
gve_tx_iovec represents a contiguous range of bytes in a registered
segment that are part of a packet to transmit. Remove the kernel virtual
address of the from the iov; it is just the segment base + iov_offset.

Saves 8 bytes per gve_tx_iovec, which saves 32 bytes per
gve_tx_buffer_state. Resolves an old TODO to shrink gve_tx_buffer_state
to fit in one cacheline.

Signed-off-by: Venkatesh Srinivas <venkateshs@google.com>
GitOrigin-RevId: 96207e16bb9851d9cb13628e69f663c0e53c6aa4