Make the CoS build output acceptable for ebuild.
The following 2 changes were required to satisfy ebuild.

1. Make tar output flat on all distros.

The -C option to tar doesn't seem to be flattening the files in the
output tarball -- files seem to be in a subdirectory inside the tar.

By changing directory before running tar, there are no folders in the
output tarball. CoS requires this.

2. Fix the path to the Makefile.

Signed-off-by: Jon Olson <jonolson@google.com>