Handle errors from gve_rx()
If the refcount of a page drops below 1 the driver needs to reset to recover
because somehow the driver no longer owns the page.

If alloc_skb fails it is because the kernel is out of memory not an internal
driver failure so a reset is not necessary.

If the descriptor has the error flag set, the error is limited to that
descriptor so a reset is not necessary.

In all cases, the counters need to continue advancing past the problematic
descriptor.

Signed-off-by: Catherine Sullivan <csully@google.com>
GitOrigin-RevId: 1c4163ddc53b3f01bed7809015c5e0c3f84f68df