gve: Roll back part of last patch
Currently, we only support the case where rx qpl's are the same length as
the rx desc ring. That may not be the case in the future so we don't want
to remove the field from describe_device and eliminate that option.

Instead, for now, we will make sure that there are enough pages per qpl
for the entire desc ring, and if there aren't we'll lower the desc
count to match.

Signed-off-by: Catherine Sullivan <csully@google.com>
GitOrigin-RevId: b2bddb8733b85b971a61c4f7d3c0c757924dd1aa