Refactor gve_rx()
There should be no functional changes here.

This pulls the buffer flip and adding frags to an skb into their own
functions.

It also restructures the code to prevent indenting 5 levels and
unreadable code as a result. Comments were added to make the code flow
and descisions being made clear.

It moves the drop due to an error to the top (we don't need to do
anything before that) and removes an info log (we really don't want to
log on every single packet).

Signed-off-by: Catherine Sullivan <csully@google.com>
GitOrigin-RevId: af0f2f647e160ebb7bfdae415a238c4a91fa3431