Fix exposing cross-reporting stats to always share stats report.
In order to expose the NIC stats to the guest even when the ethtool flag
is off, share the address and length of report at setup. When the
ethtool flag turned off, zero off the gve stats instead of detaching the
report. Only detach the report in free_stats_report.

Signed-off-by: Kuo Zhao <kuozhao@google.com>
GitOrigin-RevId: b047ac17278f18132480f72969735c9cf837afa8