gve: Fix an error handling path in 'gve_probe()'
This is a port of upstream commit 2342ae10d127:

"""
If the 'register_netdev() call fails, we must release the resources
allocated by the previous 'gve_init_priv()' call, as already done in the
remove function.

Add a new label and the missing 'gve_teardown_priv_resources()' in the
error handling path.

Fixes: 893ce44df565 ("gve: Add basic driver framework for Compute Engine Virtual NIC")
"""

Signed-off-by: David Awogbemila <awogbemila@google.com>