gve: Update README install instructions and supported features
Before this path, the GVE README was quite out of date in terms of
driver features, and installation instructions were confusing.

Remove the "manual installation" section in favor of a unified
installation section near the start of the document.

Include brief descripton of various driver features and their
limitations, and how they can be enabled and configured.

Signed-off-by: Joshua Washington <joshwash@google.com>