gve: Fix pre-page pool XDP DQ cocci patches
Before page pool was introduced, the DQ RDA buffer management strategy
was much different, following patterns similar to DQ QPL. When pagepool
was introduced, much of this logic was moved to live in coccinelle
patches. These patches were not properly updated when XDP DQ was
implemented, as it was implemented on top of a kernel that made use of
the page pool API. This led to test failures due to either non-existent
or otherwise corrupted buffers, despite the code building on all LTS
kernel versions.

For XDP buffers, the buf_size and pad fields in newly allocated
buf_states needed to be updated to packet_buffer_truesize and
rx_headroom, respectively. The buffer address in the RX buffer
descriptor similarly needed to be updated to inlude the RX pad, which
represents the XDP headroom.

XSK buffer allocation was missing altogether in the pre-pagepool
codepath, which could lead to kernel panics on XDP_REDIRECT.

This patch was tested manually on a 6.1 kernel, and on a 6.6 kernel via MTest: http://sponge2/3f1ebdc3-9772-459f-9a9f-a1a2d30ad9b7.

Change-Id: Idbec11d1df7c82511d1fd72d6c77c46adf7efaab
Signed-off-by: Joshua Washington <joshwash@google.com>
Google-Bug-Id: 439105645